requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.23.0
